+++ src/part3-webserver/breakpoints/3-make-some-room/index.ts
@@ -2,7 +2,7 @@ import { Layer, pipe } from "effect";
 import { BunRuntime } from "@effect/platform-bun";
 import * as HTTP from "./http";
 import * as WS from "./ws";
-import * as SERVER from "./server";
+import * as SERVER from "./shared";
 
 const serversLayer = Layer.merge(HTTP.Live, WS.Live);