+++ src/part3-webserver/breakpoints/3-make-some-room/ws.ts
@@ -1,5 +1,5 @@
 import { Effect, Layer } from "effect";
-import { CurrentConnections, WSSServer } from "./server";
+import { CurrentConnections, WSSServer } from "./shared";
 import * as M from "./model";
 import * as S from "@effect/schema/Schema";